Separate properties in their own file
Separate properties to avoid high memory usage on low-memory platforms.
This also adds better encoding for many properties while publishing.
